backend/requirements.txt
Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
